+++ src/PoorManFloat.h
@@ -5,20 +5,7 @@
 typedef int16_t pmf_logarithmic;
 
 #define PMF_CONST_INVALID ((pmf_logarithmic)0x8000)
-#define PMF_CONST_MAX ((pmf_logarithmic)0x7fff)
-#define PMF_CONST_1 ((pmf_logarithmic)0x0000)
-#define PMF_CONST_3_DIV_2 ((pmf_logarithmic)0x012c)
-#define PMF_CONST_128E12 ((pmf_logarithmic)0x5dba)
-#define PMF_CONST_16E6 ((pmf_logarithmic)0x2fdd)
-#define PMF_CONST_500 ((pmf_logarithmic)0x11ee)
-#define PMF_CONST_1000 ((pmf_logarithmic)0x13ee)
-#define PMF_CONST_2000 ((pmf_logarithmic)0x15ee)
-#define PMF_CONST_32000 ((pmf_logarithmic)0x1dee)
-#define PMF_CONST_16E6_DIV_SQRT_OF_2 ((pmf_logarithmic)0x2edd)
-#define PMF_CONST_21E6 ((pmf_logarithmic)0x30a5)
-#define PMF_CONST_42000 ((pmf_logarithmic)0x1eb7)
-#define PMF_CONST_21E6_DIV_SQRT_OF_2 ((pmf_logarithmic)0x2fa5)
-#define PMF_CONST_2205E11 ((pmf_logarithmic)0x5f96)
+
 pmf_logarithmic pmfl_from(uint8_t x);
 pmf_logarithmic pmfl_from(uint16_t x);
 pmf_logarithmic pmfl_from(uint32_t x);