+++ src/PoorManFloatConst.h
@@ -1,0 +1,20 @@
+#ifndef POORMANFLOATCONST_H
+#define POORMANFLOATCONST_H
+
+#include <PoorManFloat.h>
+
+#define PMF_CONST_MAX ((pmf_logarithmic)0x7fff)
+#define PMF_CONST_1 ((pmf_logarithmic)0x0000)
+#define PMF_CONST_3_DIV_2 ((pmf_logarithmic)0x012c)
+#define PMF_CONST_128E12 ((pmf_logarithmic)0x5dba)
+#define PMF_CONST_16E6 ((pmf_logarithmic)0x2fdd)
+#define PMF_CONST_500 ((pmf_logarithmic)0x11ee)
+#define PMF_CONST_1000 ((pmf_logarithmic)0x13ee)
+#define PMF_CONST_2000 ((pmf_logarithmic)0x15ee)
+#define PMF_CONST_32000 ((pmf_logarithmic)0x1dee)
+#define PMF_CONST_16E6_DIV_SQRT_OF_2 ((pmf_logarithmic)0x2edd)
+#define PMF_CONST_21E6 ((pmf_logarithmic)0x30a5)
+#define PMF_CONST_42000 ((pmf_logarithmic)0x1eb7)
+#define PMF_CONST_21E6_DIV_SQRT_OF_2 ((pmf_logarithmic)0x2fa5)
+#define PMF_CONST_2205E11 ((pmf_logarithmic)0x5f96)
+#endif