+++ CHANGELOG.md
@@ -5,6 +5,9 @@ TODO:
 - rename RampConstAcceleration to e.g. RampControl
 - for esp-idf 5 make use of espressif resource management of rmt channels
 
+pre-0.31.2:
+- Move constants out of PoorManFloat.h
+
 0.31.1:
 - Fix for issue #280: stopMove() is interrupted if followed by update of speed/acceleration
 - Related to issue #280: `isStopping()` did not work as expected and this should be fixed, too.